docs/extraction-draft/atproto-oauth-py/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "atproto-oauth-py"
version = "0.1.0"
description = "AT Protocol OAuth 2.1 + DPoP library and CLI: handle/DID/PDS resolution, PAR, PKCE, refresh-token flow, and DPoP-bound resource calls."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [{ name = "tenorune" }]
keywords = ["atproto", "bluesky", "oauth", "oauth2", "dpop", "pkce"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Internet",
  "Topic :: Security :: Cryptography",
]
dependencies = [
  "httpx>=0.27",
  "pyjwt[crypto]>=2.8",
  "cryptography>=42",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "respx>=0.21",
]

[project.urls]
Homepage = "https://github.com/tenorune/atproto-oauth-py"
Issues = "https://github.com/tenorune/atproto-oauth-py/issues"

[project.scripts]
atproto-oauth = "atproto_oauth.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/atproto_oauth"]
